Fix the clippy warnings a cold build surfaces
Warm clippy runs replay cached MIR-lint results, so these only showed up
on a from-clean build:

- Elide nine clones in Task::perform result mappers (iced 0.14 made the
  mapper FnOnce, so the captured key can move instead of cloning per
  fetch) and drop the redundant helper clones the old Fn bound required
- animated_image::Frames::from_bytes returns Option instead of
  Result<_, ()>; the unit error carried nothing
- Assign-op in the video clip-host translation walk, no-op f32 cast in
  the scale-factor clamp, while-let drain in the room-list watcher

+++ crates/app/src/main.rs
@@ -72,7 +72,7 @@ fn main() -> iced::Result {
         // iced (it would divide the viewport into a degenerate size).
         .scale_factor(|state: &ui::App| {
             let scale = state.theme.ui_scale;
-            if scale.is_finite() { scale.clamp(0.8, 1.5) as f32 } else { 1.0 }
+            if scale.is_finite() { scale.clamp(0.8, 1.5) } else { 1.0 }
         })
         .default_font(default_font)
         // cosmic-text's glyph-fallback walk (default font -> per-script
